+++ server/src/test/kotlin/com/adobe/testing/s3mock/store/MultipartStoreTest.kt
@@ -211,6 +211,7 @@ internal class MultipartStoreTest : StoreTestBase() {
         multipartUploadInfo,
         "location",
         NO_CHECKSUM,
+        NO_CHECKSUMTYPE,
         NO_CHECKSUM_ALGORITHM,
       )
     val md5 = MessageDigest.getInstance("MD5")
@@ -288,6 +289,7 @@ internal class MultipartStoreTest : StoreTestBase() {
       multipartUploadInfo,
       "location",
       NO_CHECKSUM,
+      NO_CHECKSUMTYPE,
       NO_CHECKSUM_ALGORITHM,
     )
 
@@ -355,6 +357,7 @@ internal class MultipartStoreTest : StoreTestBase() {
       multipartUploadInfo,
       "location",
       NO_CHECKSUM,
+      NO_CHECKSUMTYPE,
       NO_CHECKSUM_ALGORITHM,
     )
 
@@ -445,6 +448,7 @@ internal class MultipartStoreTest : StoreTestBase() {
       multipartUploadInfo,
       "location",
       checksum,
+      ChecksumType.COMPOSITE,
       ChecksumAlgorithm.CRC32,
     )
   }
@@ -504,6 +508,7 @@ internal class MultipartStoreTest : StoreTestBase() {
         multipartUploadInfo,
         "location",
         wrongOverallChecksum,
+        ChecksumType.COMPOSITE,
         ChecksumAlgorithm.CRC32,
       )
     }.isInstanceOf(S3Exception::class.java)
@@ -570,6 +575,7 @@ internal class MultipartStoreTest : StoreTestBase() {
         multipartUploadInfo,
         "location",
         checksum,
+        ChecksumType.COMPOSITE,
         ChecksumAlgorithm.CRC32,
       )
     }.isInstanceOf(S3Exception::class.java)
@@ -704,6 +710,7 @@ internal class MultipartStoreTest : StoreTestBase() {
       multipartUploadInfo,
       "location",
       NO_CHECKSUM,
+      NO_CHECKSUMTYPE,
       NO_CHECKSUM_ALGORITHM,
     )
 
@@ -757,6 +764,7 @@ internal class MultipartStoreTest : StoreTestBase() {
       multipartUploadInfo,
       "location",
       NO_CHECKSUM,
+      NO_CHECKSUMTYPE,
       NO_CHECKSUM_ALGORITHM,
     )
 
@@ -842,6 +850,7 @@ internal class MultipartStoreTest : StoreTestBase() {
       multipartUploadInfo1,
       "location",
       NO_CHECKSUM,
+      NO_CHECKSUMTYPE,
       NO_CHECKSUM_ALGORITHM,
     )
     multipartStore.completeMultipartUpload(
@@ -854,6 +863,7 @@ internal class MultipartStoreTest : StoreTestBase() {
       multipartUploadInfo2,
       "location",
       NO_CHECKSUM,
+      NO_CHECKSUMTYPE,
       NO_CHECKSUM_ALGORITHM,
     )
 
@@ -1125,6 +1135,7 @@ internal class MultipartStoreTest : StoreTestBase() {
       multipartUploadInfo,
       "location",
       NO_CHECKSUM,
+      NO_CHECKSUMTYPE,
       NO_CHECKSUM_ALGORITHM,
     )
     val s = objectStore.getS3ObjectMetadata(bucket, id, null)!!.dataPath