+++ CHANGELOG.md
@@ -151,6 +151,7 @@ Version 5.x is JDK17 LTS bytecode compatible, with Docker and JUnit / direct Jav
 * Features and fixes
   * Get object with range now returns the same headers as non-range calls.
   * Docker: Copy "s3mock.jar" to "/opt/", run with absolute path reference to avoid issues when working directory is changed. (fixes #2827)
+  * S3Mock supports ChecksumType.FULL_OBJECT for Multipart uploads (fixes #2843)
 * Refactorings
   * Use Jackson 3 annotations and mappers.
   * AWS has deprecated SDK for Java v1 and will remove support EOY 2025.